Force float64 in streaming combined writer to avoid schema drift
When build_combined_file streams ticker files into the parquet writer,
the first chunk would set 'transactions' as int64 (when integer values
are present) but later chunks where the column is all-NaN would arrive
as float64, breaking the writer's strict schema check.

Explicit .astype("float64") on each chunk's float cols guarantees a
stable schema across all writes.

+++ src/data/fetch_prices.py
@@ -220,8 +220,11 @@ def build_combined_file():
                     df[col] = pd.NA
             df = df[expected_cols]
             df["date"] = pd.to_datetime(df["date"])
+            # Force float64 (not whatever pd.to_numeric infers) so chunks
+            # where transactions/vwap are present as ints don't break the
+            # writer when later chunks fill those columns with NaN.
             for col in float_cols:
-                df[col] = pd.to_numeric(df[col], errors="coerce")
+                df[col] = pd.to_numeric(df[col], errors="coerce").astype("float64")
 
             table = pa.Table.from_pandas(df, preserve_index=False)
             if writer is None:
